bind: ensure enum.Enum is imported in generated code as needed
This fix removes the static import of enum.Enum that was added only to the exe preamble, and replaces it with a conditional import added to all generated packages, only if there is at least one enum to generate.
The conditional import improves python2 legacy support where Enum is an external dependency that must be present.